requirements.txt
# Core Dependencies
hikari>=2.3.3
hikari-arc>=1.6.0

# AI Integration
google-generativeai>=0.8.0

# Web Scraping & HTTP
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
aiosqlite>=0.19.0

# Environment Management
python-dotenv>=1.0.0

# Date & Time Handling
pytz>=2024.1

# Utilities
typing-extensions>=4.8.0

# Development Dependencies (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# mypy>=1.5.0
